pyproject.toml
[project]
name = "charging-prices"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fire>=0.7.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "requests-cache>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.9",
]
